Fixes #84 Add withSink param
graph.Terminate() has a new parameter, withSink, which will skip adding the sink at the end of a command graph when it is set to false.
It will be false if the command has JustPrefix=true, it means it is a parent command under command that can not be executed.